fix: mf-5271 warming up persist caching
