+++ packages/mask/package.json
@@ -93,7 +93,9 @@
     "@servie/events": "^3.0.0",
     "@sinonjs/text-encoding": "^0.7.2",
     "@splinetool/react-spline": "2.2.6",
+    "@tanstack/query-sync-storage-persister": "^4.35.0",
     "@tanstack/react-query-devtools": "^4.29.25",
+    "@tanstack/react-query-persist-client": "^4.35.0",
     "@types/bn.js": "^4.11.6",
     "@types/d3": "5.16.4",
     "@types/elliptic": "^6.4.14",
