+++ packages/shared/src/UI/contexts/index.ts
@@ -3,3 +3,4 @@ export * from './components/index.js'
 export * from './SharedContextProvider.js'
 export * from './PageUIProvider.js'
 export * from './SiteUIProvider.js'
+export * from './localStoragePersist.js'
